Disable "Invoke Tests & Publish Results" job
The condition for running the "Invoke Tests & Publish Results" job in the `release-pipeline.yml` file has been modified. Previously, the job would run if the `NewVersion` job succeeded and the `validVersion` output was `True`. The new change comments out the original condition and replaces it with a condition that always evaluates to `false`, effectively preventing the job from running.