Update job dependencies in CI pipeline
The `if` condition for the `PublishNugetPackages` job has been commented out, causing it to always run regardless of the `InvokeTests` job result.

The `if` condition for the `NewRelease` job has been modified to remove the dependency on the `InvokeTests` job's result. Now, the `NewRelease` job will run if the `NewVersion` job is successful, the new version is valid, and the `PublishNugetPackages` job is successful.

The `InvokeTests` job has been removed from the `needs` section of the `NewRelease` job, indicating that the `NewRelease` job no longer depends on the `InvokeTests` job.