Make PublishNugetPackages job run unconditionally
Modified the condition for running the `PublishNugetPackages` job in the `release-pipeline.yml` file. The job will now always run, regardless of the outcome of the `InvokeTests` job, by using `${{ always() }}`. This change ensures that NuGet packages are published and pushed in every pipeline run, enhancing the robustness of the pipeline.