Simplify NewRelease job condition in release pipeline
The conditional statement for the `NewRelease` job in the `release-pipeline.yml` file has been simplified. Previously, the job would only run if the `NewVersion` job was successful, the `NewVersion` output `validVersion` was `True`, and the `PublishNugetPackages` job was successful. Now, the job will run if only the `PublishNugetPackages` job is successful.