Make NewRelease job always run in release-pipeline.yml
The condition for running the NewRelease job in the release-pipeline.yml file has been modified. Previously, the job would run only if the PublishNugetPackages job succeeded (if: ${{ needs.PublishNugetPackages.result == 'success' }}). The updated condition if: ${{ always() }} ensures that the NewRelease job will always run, regardless of the outcome of the PublishNugetPackages job.